refactor: RealmManager 구조 정리
